moved Admin actions from actions.js to actions/AdminActions.js and renamed slideshow to photos in Photos.js
