+++ bc-react/src/routes/Photos.js
@@ -6,7 +6,7 @@ import Header from '../components/Header';
 import MessageBoardToggle from '../components/MessageBoardToggle';
 
 
-class Slideshow extends Component {
+class Photos extends Component {
 
   prev(e) {
       e.preventDefault();
@@ -58,4 +58,4 @@ class Slideshow extends Component {
       );
     }
 }
-export default Slideshow;
+export default Photos;
